+++ tests/Model/YamlIndent/YamlIndentCheckerTest.php
@@ -26,6 +26,8 @@ class YamlIndentCheckerTest extends TestCase
     {
         $standardParametersData = $this->getStandardsParametersData(YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT);
         $pathToFiles = [
+            __DIR__ . '/resource/unfixed/arraysWithUnquotedColons.yml',
+            __DIR__ . '/resource/unfixed/kubernetes-postgres.yml',
             __DIR__ . '/resource/unfixed/kustomization.yaml',
             __DIR__ . '/resource/unfixed/shopsys-service.yml',
             __DIR__ . '/resource/unfixed/symfony-config.yml',
@@ -69,6 +71,8 @@ class YamlIndentCheckerTest extends TestCase
     {
         $standardParametersData = $this->getStandardsParametersData(YamlStandardConfigDefinition::CONFIG_PARAMETERS_INDENTS_COMMENTS_WITHOUT_PARENT_VALUE_DEFAULT);
         $pathToFiles = [
+            __DIR__ . '/resource/fixed/arraysWithUnquotedColons.yml',
+            __DIR__ . '/resource/fixed/kubernetes-postgres.yml',
             __DIR__ . '/resource/fixed/kustomization.yaml',
             __DIR__ . '/resource/fixed/shopsys-service.yml',
             __DIR__ . '/resource/fixed/symfony-config.yml',